Fixing administrator -> member -> edit
When editing a member in the administrator area the department and roles drop down were returning empty due to either, the master member's department being 0 whilst the team department was set and being used for display but not edit or the role was set to 0 and subsequently was not looking up it's value correctly. This fixes both.